+++ src/data.cpp
@@ -163,29 +163,29 @@ void Data::mapPreprocessBedFile(const string &preprocessedBedFile)
 {
     // Calculate the expected file sizes - cast to size_t so that we don't overflow the unsigned int's
     // that we would otherwise get as intermediate variables!
-    const size_t ppBedSize = size_t(numInds) * size_t(numIncdSnps) * sizeof(float);
+    const size_t ppBedSize = size_t(numInds) * size_t(numIncdSnps) * sizeof(double);
 
     // Open and mmap the preprocessed bed file
     ppBedFd = open(preprocessedBedFile.c_str(), O_RDONLY);
     if (ppBedFd == -1)
         throw("Error: Failed to open preprocessed bed file [" + preprocessedBedFile + "]");
 
-    ppBedMap = reinterpret_cast<float *>(mmap(nullptr, ppBedSize, PROT_READ, MAP_SHARED, ppBedFd, 0));
+    ppBedMap = reinterpret_cast<double *>(mmap(nullptr, ppBedSize, PROT_READ, MAP_SHARED, ppBedFd, 0));
     if (ppBedMap == MAP_FAILED)
         throw("Error: Failed to mmap preprocessed bed file");
 
     // Now that the raw data is available, wrap it into the mapped Eigen types using the
     // placement new operator.
     // See https://eigen.tuxfamily.org/dox/group__TutorialMapClass.html#TutorialMapPlacementNew
-    new (&mappedZ) Map<MatrixXf>(ppBedMap, numInds, numIncdSnps);
+    new (&mappedZ) Map<MatrixXd>(ppBedMap, numInds, numIncdSnps);
 }
 
 void Data::unmapPreprocessedBedFile()
 {
     // Unmap the data from the Eigen accessors
-    new (&mappedZ) Map<MatrixXf>(nullptr, 1, 1);
+    new (&mappedZ) Map<MatrixXd>(nullptr, 1, 1);
 
-    const auto ppBedSize = numInds * numIncdSnps * sizeof(float);
+    const auto ppBedSize = numInds * numIncdSnps * sizeof(double);
     munmap(ppBedMap, ppBedSize);
     close(ppBedFd);
 }
@@ -210,7 +210,7 @@ void Data::mapCompressedPreprocessBedFile(const string &preprocessedBedFile,
     if (ppBedFd == -1)
         throw("Error: Failed to open preprocessed bed file [" + preprocessedBedFile + "]");
 
-    ppBedMap = reinterpret_cast<float *>(mmap(nullptr, ppBedSize, PROT_READ, MAP_SHARED, ppBedFd, 0));
+    ppBedMap = reinterpret_cast<double *>(mmap(nullptr, ppBedSize, PROT_READ, MAP_SHARED, ppBedFd, 0));
     if (ppBedMap == MAP_FAILED)
         throw("Error: Failed to mmap preprocessed bed file");
 }
